+++ src/main/java/com/example/dpm/repository/MemberRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.dpm.model.MemberEntity;
 
-public interface MemberRepository extends JpaRepository<MemberEntity, Integer>{
+public interface MemberRepository extends JpaRepository<MemberEntity, Long>{
 
-	 Optional<MemberEntity> findById(int id);
+	 Optional<MemberEntity> findById(Long id);
 
 	 Optional<MemberEntity> findByRefreshToken(String refreshToken);
 }
